01_kafka/01_producer/requirements.txt
confluent-kafka==2.8.0
python-dotenv